minor optimisations
